projeto_hotel/requirements.txt
gspread
oauth2client
openai
requests
Faker
python-dotenv